% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{create_table_lables}
\alias{create_table_lables}
\title{create_table_lables Datensatzes mit Vauluelabel ausstatten}
\usage{
create_table_lables(table)
}
\arguments{
\item{table}{data.frame aus get_mean_data}
}
\value{
data_with_label = Datensatz mit Valuelabels
}
\description{
create_table_lables soll spezifische Variablen eines Datensatzes mit Vauluelabel ausstatten
}
\examples{
      create_table_lables(table = data)
}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{data_with_label}
