% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{expand_table}
\alias{expand_table}
\title{expand_table}
\usage{
expand_table(table, table_type)
}
\description{
expand_table add empty rows if variable in year is empty.
}
\examples{
expand_table(
  table = protected_table, grouping_variable_one = grouping_variable_one,
  grouping_variable_two = grouping_variable_two,
  grouping_variable_three = grouping_variable_three,
  grouping_count = grouping_count, table_type = "prop"
)

}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
