% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{get_data}
\alias{get_data}
\title{get_data erzeugt Teilmenge eines Datensatzes}
\usage{
get_data(
  datasetnum,
  datasetfac,
  variable,
  year,
  weight,
  diffcount,
  diffvars,
  vallabel
)
}
\arguments{
\item{datasetnum}{data.frame mit numeric Variablen (z.B. plattform_data)}

\item{datasetfac}{data.frame mit factor Variablen (z.B. plattform_data)}

\item{variable}{Name Analysevariable als string (z.B. "pglabnet" )}

\item{year}{Name Erhebungsjahrvariable als string (z.B. "syear")}

\item{weight}{Name Gewichtsvariable als string (z.B. "phrf")}

\item{diffcount}{Anzahl gewuenschter Differenzierungen (z.B. 2) (range 0-3)}

\item{diffvars}{Vector mit Differenzierungsvariablen (z.B. c("alter_gr", "sex", "Bildungsniveau")) (maximal 3 Variablen)}

\item{vallabel}{Valuelabel sollen verwendet werden (z.B: vallabel = TRUE) (TRUE/FALSE)}
}
\value{
variable.values.valid ist ein Datensatz mit validen Werten der
Variablen (variable, year, weight, diffvars)
}
\description{
get_data soll Ausgangs-Datensatz im long Format
auf bestimmte Variablen beschraenken (variable, year, weight, diffvars) und
ausschliesslich valide Werte enthalten.
}
\examples{
      get_data(datasetnum =  data.file.num, 
      datasetfac = data.file.fac,
      variable = "pglabnet", 
      year = "syear", 
      weight = "phrf",
      diffcount = diffcount,
      diffvars = c("alter_gr", "sex", "Bildungsniveau"),
      vallabel = TRUE)
}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{get_data}
