% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{get_mean_values}
\alias{get_mean_values}
\title{get_mean_values erstellt Mittelwert/Mediantabellen mit Mittelwert-Konfidenzinterval}
\usage{
get_mean_values(dataset, year, diffcount, diffvar1, diffvar2, diffvar3)
}
\arguments{
\item{dataset}{data.frame aus get_data (z.B. plattform_data)}

\item{year}{Jahresvariable als string (z.B. "year")}

\item{diffcount}{Anzahl gewünschter Differenzierungen (z.B. 2) (range 0-3)}

\item{diffvar1}{Name Differenzierungsvariable 1 als string ("" moeglich)}

\item{diffvar2}{Name Differenzierungsvariable 2 als string ("" moeglich)}

\item{diffvar3}{Anzahl Differenzierungsvariable 3 als string ("" moeglich)}
}
\value{
mean.values = Datensatz mit n, weighted_mean, weighted_median, Konfidenzintervalle
}
\description{
get_mean_values soll gewichtete Mittelwert/Mediantabellen mit Mittelwert-Konfidenzinterval
erzeugen mit den Informationen n = Groesse der Subgruppe, weighted_mean = gewichteter
Mittelwert, weighted_median = gewichteter Median, lower_ci = unteres Konfidenzintervall 95\%,
upper_ci = oberes Konfidenzintervall 95\%
}
\examples{
      get_mean_values(dataset = data, 
                      year = "year", 
                      diffcount = 2,
                      diffvar1 = "sex",
                      diffvar2 = "alter_gr",
                      diffvar3 = "")
}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{get_mean_values}
