% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{get_prop_values}
\alias{get_prop_values}
\title{function get_prop_values shall create weighted proportion tables with
confidence intervals}
\usage{
get_prop_values(dataset, groupvars, alpha)
}
\arguments{
\item{dataset}{data.frame from get_data (e.g. platform_data)}

\item{groupvars}{vector with all variables in the dataset
(e.g. c("usedvariable", "year", "sex"))}

\item{alpha}{Alpha for setting the confidence interval (e.g. 0.05)}
}
\value{
data_prop_complete_ci = data set with n, percent, lower_confidence,
upper_confidence
}
\description{
get_mean_values shall create weighted proportion value tables
with confidence intervals
create with the information n = size of the subgroup, percent = weighted
proportion value, lower_confidence = lower confidence interval,
upper_confidence = upper 95\% confidence interval
}
\examples{
get_prop_values(
  dataset = data,
  groupvars = c("usedvariable", "year", "sex"),
  alpha = 0.05
)

}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{get_prop_values}
