% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{get_table_export}
\alias{get_table_export}
\title{get_table_export Export von Mittelwert-oder Anteilswertswert-Tabellen}
\usage{
get_table_export(
  table,
  variable,
  metadatapath,
  exportpath,
  diffcount,
  tabletype
)
}
\arguments{
\item{table}{produzierter data.frame aus get_protected_values (z.B. plattform_data)}

\item{variable}{Name Analysevariable aus den Rohdaten als string ("pglabnet")}

\item{metadatapath}{Pfad zu den Metadaten mit Variablennamen und Tabellentyp im Datensatz als string}

\item{exportpath}{Exportordner als string}

\item{diffcount}{Anzahl an Differenzierungen als numeric (0-3 eraubt)}

\item{tabletype}{Art der zu verarbeitenden Tabelle ("mean" oder "prop")}
}
\value{
data.csv = exportierte Tabelle als csv
}
\description{
get_table_export soll erzeugte Mittelwert-oder Anteilswertswert-Tabellen als csv Exportieren
}
\examples{
       get_table_export(table = protected.table, variable = "usedvariable", 
                        metadatapath = paste0(metapath, "varnames.csv"),
                        exportpath = exportpath, diffcount = 2, 
                        tabletype = "mean")
}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{data.csv}
