% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{json_create}
\alias{json_create}
\title{json_create}
\usage{
json_create(table, variable, table_type)
}
\description{
json_create creates json metadata
}
\examples{
#

}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
