% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{loadpackage}
\alias{loadpackage}
\title{loadpackage laedt benoetigte R-Pakete}
\usage{
loadpackage(x)
}
\arguments{
\item{x}{Namen der benoetigten R-Packages als Vector (z.B. c("foreign", "dplyr", "tidyverse", "readstata13"))}
}
\description{
loadpackage soll benoetigte Pakete, wenn noetig installieren und in library laden
}
\examples{
      loadpackage(c("foreign", "dplyr", "tidyverse", "readstata13")) 
}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{loadpackage}
