% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{create_table_lables}
\alias{create_table_lables}
\title{create_table_lables data set with vauluelabel}
\usage{
create_table_lables(table, grouping_variables)
}
\arguments{
\item{table}{data.frame from get_mean_data}
}
\value{
data_with_label = data set with value labels
}
\description{
create_table_lables is to provide specific variables of a
dataset with vauluelabel
}
\examples{
create_table_lables(table = data)

}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{data_with_label}
