% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{get_data}
\alias{get_data}
\title{get_data creates subset of a data set}
\usage{
get_data(variable, grouping_variables, value_label)
}
\arguments{
\item{variable}{name analysis variable as string (e.g. "pglabnet" )}

\item{grouping_variables}{Vector with differentiation variables
(e.g. c("age_gr", "sex", "education level")) (maximum 3 variables)}

\item{value_label}{Valuelabel should be used (e.g.: value_label = TRUE)
(TRUE/FALSE)}

\item{datafile_without_labels}{data.frame with numeric variables
(e.g. platform_data)}

\item{datafile_with_labels}{data.frame with factor variables
(e.g. platform_data)}

\item{year}{Name survey year variable as string (e.g. "syear" )}

\item{weight}{Name weight variable as string (e.g. "phrf")}

\item{grouping_count}{Number of desired differentiations (e.g. 2)
(range 0-3)}
}
\value{
variable.values.valid is a data set with valid values of the
variables (variable, year, weight, grouping_variables)
}
\description{
get_data to get the output dataset in long format
limited to certain variables (variable, year, weight, grouping_variables) and
contain only valid values.
}
\examples{
get_data(
  datafile_without_labels = datafile_without_labels,
  datafile_with_labels = datafile_with_labels,
  variable = "pglabnet",
  year = "syear",
  weight = "phrf",
  grouping_count = grouping_count,
  grouping_variables = c("alter_gr", "sex", "Bildungsniveau"),
  value_label = TRUE
)

}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{get_data}
