% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{get_mean_values}
\alias{get_mean_values}
\title{get_mean_values creates mean/median tables with mean, median, n,
percentiles, confidence interval}
\usage{
get_mean_values(dataset, grouping_variables)
}
\arguments{
\item{dataset}{data.frame from get_data (e.g. platform_data)}

\item{year}{year variable as string (e.g. "year")}

\item{grouping_count}{number of desired differentiations (e.g. 2) (range 0-3)}

\item{grouping_variable_one}{name differentiation variable 1 as string or ""}

\item{grouping_variable_two}{name differentiation variable 2 as string or ""}

\item{grouping_variable_three}{Number of differentiation variable 3 as string
("" possible)}
}
\value{
data = dataset with mean, median, n, percentiles, confidence interval
}
\description{
get_mean_values creates weighted mean/median tables with
n, percentiles, confidence interval
}
\examples{
get_mean_values(
  dataset = data,
  year = "year",
  grouping_count = 2,
  grouping_variable_one = "sex",
  grouping_variable_two = "alter_gr",
  grouping_variable_three = ""
)

}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{get_mean_values}
