% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{get_prop_values}
\alias{get_prop_values}
\title{Funktion get_prop_values soll gewichtete Anteilswert-Tabellen mit Konfidenzintervallen erstellen}
\usage{
get_prop_values(dataset, groupvars, alpha)
}
\arguments{
\item{dataset}{data.frame aus get_data (z.B. plattform_data)}

\item{groupvars}{Vector mit allen Variablen im Datensatz (z.B. c("usedvariable", "year", "sex"))}

\item{alpha}{Alpha fuer die Festlegung der Konfidenzintervalles (z.B. 0.05)}
}
\value{
data_prop_complete_ci = Datensatz mit n, weighted_percent, lower_ci, upper_ci
}
\description{
get_mean_values soll gewichtete Anteilswert-Tabellen mit Konfidenzintervallen erstellen
erzeugen mit den Informationen n = Groesse der Subgruppe, weighted_percent = gewichteter
Anteilswert, lower_ci = unteres Konfidenzintervall, upper_ci = oberes Konfidenzintervall 95\%
}
\examples{
      get_prop_values(dataset = data, 
                      groupvars = c("usedvariable", "year", "sex"), 
                      alpha = 0.05)
}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{get_prop_values}
