% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{get_protected_values}
\alias{get_protected_values}
\title{get_protected_values soll bestimmte Zelleninhalte entfernen}
\usage{
get_protected_values(dataset, cell.size)
}
\arguments{
\item{dataset}{data.frame aus get_prop_values oder get_mean_table}

\item{cell.size}{maximal zulaessige Zellengroesse (z.B. 30)}
}
\value{
protected.data (Datensatz mit n, weighted_mean/weighted_percent, weighted_median, n, Konfidenzintervalle ausschlie?lich mit Zellen >= cell.size)
}
\description{
get_protected_values soll Zelleninhalte von gewichteten Anteilswert-Tabellen oder Mittelwert-Tabelle
loeschen wenn eine Minimalbesetzung unterschritten wird.
}
\examples{
      get_prop_values(dataset = data, alpha = 0.05)
}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{get_prop_values}
