% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_creation_FUN.R
\name{get_table_export}
\alias{get_table_export}
\title{get_table_export Export of mean value or proportion value tables}
\usage{
get_table_export(table, variable, table_type)
}
\arguments{
\item{table}{produced data.frame from get_protected_values
(e.g. platform_data)}

\item{variable}{name analysis variable from raw data as string ("pglabnet")}

\item{table_type}{Type of table to be processed ("mean" or "prop")}

\item{metadata_path}{Path to the metadata with variable name and table type
in the dataset as string}

\item{export_path}{export folder as string}

\item{grouping_count}{number of differentiations as numeric (0-3 robbed)}
}
\value{
data_csv = exportierte Tabelle als csv
}
\description{
get_table_export export created mean or proportion tables as csv
}
\examples{
get_table_export(
  table = protected_table, variable = "usedvariable",
  metadata_path = paste0(metadata_path, "varnames.csv"),
  export_path = export_path, grouping_count = 2,
  table_type = "mean"
)

}
\author{
Stefan Zimmermann, \email{szimmermann@diw.de}
}
\keyword{data_csv}
